+++ src/data/mod.rs
@@ -1,0 +1,8 @@
+// Copyright (C) 2020 Daniel Mueller <deso@posteo.net>
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+/// Definitions for the first version of the Alpaca Data API.
+pub mod v1;
+
+/// The API base URL used for retrieving market data.
+pub(crate) const DATA_BASE_URL: &str = "https://data.alpaca.markets";