+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added support for accessing `/v2/account/configurations` endpoint
+- Added support for querying `/v1/bars/<timeframe>` endpoint
 - Switched from using `log` to `tracing` as a logging/tracing provider
 - Switched to using `serde_urlencoded` for encoding query parameters
 - Bumped `http-endpoint` dependency to `0.1.1`